organize test dir
