+++ smart-contracts/contracts/cl-vault/src/tests/e2e/mod.rs
@@ -1,0 +1,2 @@
+mod deposit_withdraw;
+mod proptest;
