+++ smart-contracts/contracts/cl-vault/src/tests/e2e/proptest.rs
