+++ smart-contracts/contracts/cl-vault/src/tests/integration/mod.rs
@@ -1,0 +1,2 @@
+mod multirange_deposit;
+mod range;
